TS => MJS: fsc
Migrate fjs/fsc/module.f.ts to module.f.mjs under the stage-1
TypeScript-to-mjs migration, folded into the same PR since it became
a fresh dependency leaf once text/ascii migrated. All other runtime
deps (function/operator, range_map, list, function, range, asserts)
were already .f.mjs.

The generic helpers (union, reduce, rangeSet, create) use the same
per-arrow @template/@param/@returns style as sorted_list/range_map.
Only `terminal`, `init`, and the module-embedded `proof` are exported;
declaration emit checked clean (0 elided).

proof.f.ts stays TypeScript: it still imports the unmigrated
media/json module; only the module import specifier is updated.

Re-anchor two todo docs whose line numbers and quoted current-source
snippets referenced the pre-migration file.
