+++ CHANGELOG.md
@@ -20,10 +20,10 @@ history.
 
 ## Unreleased
 
-- **BREAKING CHANGES:** `fjs/ci/config` and `fjs/text/ascii` migrate from
-  authored TypeScript (`.f.ts`) to JSDoc-typed JavaScript (`.f.mjs`) under
-  the stage-1 TypeScript-to-mjs migration — importers must use the
-  `.f.mjs` specifier
+- **BREAKING CHANGES:** `fjs/ci/config`, `fjs/text/ascii`, and `fjs/fsc`
+  migrate from authored TypeScript (`.f.ts`) to JSDoc-typed JavaScript
+  (`.f.mjs`) under the stage-1 TypeScript-to-mjs migration — importers
+  must use the `.f.mjs` specifier
   [#1479](https://github.com/functionalscript/functionalscript/pull/1479)
 - **BREAKING CHANGES:** `fjs/types/patricia_trie`, `fjs/types/sorted_list`,
   `fjs/types/range_map`, `fjs/types/range_set`, `fjs/crypto/pow`,
