--- fjs/fsc/module.f.ts
@@ -1,111 +1,0 @@
-/**
- * FunctionalScript command utilities for compile workflows.
- *
- * @module
- */
-import { strictEqual } from '../types/function/operator/module.f.mjs'
-import {
-    merge as rangeMapMerge,
-    fromRange,
-    get,
-    type RangeMapArray,
-    type RangeMerge,
-} from '../types/range_map/module.f.mjs'
-import { reduce as listReduce, toArray, map, type List } from '../types/list/module.f.mjs'
-import { range as asciiRange } from '../text/ascii/module.f.mjs'
-import { flip, fn } from '../types/function/module.f.mjs'
-import { one, type Range } from '../types/range/module.f.mjs'
-import { assertEq } from '../asserts/module.f.mjs'
-
-const fromCharCode = String.fromCharCode
-
-type Result = readonly[readonly string[], ToResult]
-
-type ToResult = (codePoint: number) => Result
-
-type CreateToResult<T> = (state: T) => ToResult
-
-type State<T> = RangeMapArray<CreateToResult<T>>
-
-const unexpectedSymbol: ToResult
-    = codePoint => [[`unexpected symbol ${codePoint}`], unexpectedSymbol]
-
-const def: <T>(state: T) => ToResult
-    = () => unexpectedSymbol
-
-const union = <T>(a: CreateToResult<T>) => (b: CreateToResult<T>): CreateToResult<T> => {
-    if (a === def || a === b) { return b }
-    if (b === def) { return a }
-    throw [a, b]
-}
-
-const empty: readonly never[] = []
-
-const reduce = <T>(a: List<State<T>>): State<T> => {
-    const merge: RangeMerge<CreateToResult<T>>
-        = rangeMapMerge({
-            union,
-            equal: strictEqual,
-            def,
-        })
-    return toArray(listReduce(merge)(empty)(a))
-}
-
-const codePointRange = flip(fromRange(def))
-
-const range = fn(asciiRange).map(codePointRange).result
-
-const rangeSet = (l: readonly string[]) => <T>(f: CreateToResult<T>): State<T> => {
-
-    const codePointRange: (a: Range) => (f: CreateToResult<T>) => State<T>
-        = flip(fromRange(def))
-
-    const g: (r: string) => State<T>
-        = r => codePointRange(asciiRange(r))(f)
-
-    return reduce(map(g)(l))
-}
-
-const create = <T>(a: List<State<T>>): CreateToResult<T> => {
-    const i = reduce(a)
-    const x: (i: State<T>) => (v: number) => (v: T) => ToResult
-        = get(def)
-    return v => c => x(i)(c)(v)(c)
-}
-
-export const terminal = -1
-
-const toInit: () => ToResult
-    = () => () => [[], init]
-
-const single = (c: string): State<undefined> =>
-    range(c)(() => () => [[c], unexpectedSymbol])
-
-const punctuation = "!\"%&'()*+,-./:;<=>?[]^`{|}~" as const
-
-export const init: ToResult = create([
-    codePointRange(one(terminal))(toInit),
-    rangeSet(['\t', ' ', '\n', '\r'])(toInit),
-    rangeSet(['$', '_', 'AZ', 'az'])(() => c => [[fromCharCode(c)], unexpectedSymbol]),
-    range('09')(() => a => [[fromCharCode(a)], unexpectedSymbol]),
-    ...[...punctuation].map(single),
-])(undefined)
-
-export const proof = {
-    // union throws when two distinct non-def handlers are merged for the same range;
-    // this path is unreachable through the public API (init has no overlapping ranges),
-    // so we exercise it here where the private union function is in scope.
-    throw: {
-        unionConflict: () => {
-            const a: CreateToResult<undefined> = _s => unexpectedSymbol
-            const b: CreateToResult<undefined> = _s => unexpectedSymbol
-            a(undefined)
-            b(undefined)
-            union(a)(b)
-        }
-    },
-    // `def` is the range-map's default handler; the public API never calls it directly
-    // (`init` covers every code point), so exercise it here where it's in scope.
-    defHandler: () =>
-        assertEq(def(undefined), unexpectedSymbol)
-}
