+++ fjs/js/todo/174.md
@@ -3,21 +3,29 @@
 **Priority:** P3
 **Status:** open
 
-`fjs/fsc/module.f.ts` and `fjs/js/tokenizer/module.f.ts` are the only two
+`fjs/fsc/module.f.mjs` and `fjs/js/tokenizer/module.f.ts` are the only two
 code-point scanners in the tree, and both hand-roll the *same* Mealy machine
 over `range_map`: a state is a continuation `(codePoint) => [output, nextState]`,
 transitions are looked up in a `RangeMapArray` of continuations, and overlapping
 ranges are merged with one identical conflict rule. They differ only in the
 output payload type.
 
-```ts
-// fjs/fsc/module.f.ts:35
-const union = <T>(a: CreateToResult<T>) => (b: CreateToResult<T>): CreateToResult<T> => {
-    if (a === def || a === b) { return b }
-    if (b === def) { return a }
-    throw [a, b]
-}
+```js
+// fjs/fsc/module.f.mjs:33
+const union =
+    /**
+     * @template T
+     * @param {_CreateToResult<T>} a
+     * @returns {(b: _CreateToResult<T>) => _CreateToResult<T>}
+     */
+    a => b => {
+        if (a === def || a === b) { return b }
+        if (b === def) { return a }
+        throw [a, b]
+    }
+```
 
+```ts
 // fjs/js/tokenizer/module.f.ts:330
 const union
     : <T>(def: CreateToToken<T>) => (a: CreateToToken<T>) => (b: CreateToToken<T>) => CreateToToken<T>
@@ -30,14 +38,14 @@ const union
 
 The rest of the machinery is the same algorithm with the payload type swapped:
 
-| concern | `fsc` (`module.f.ts`) | `js/tokenizer` (`module.f.ts`) |
+| concern | `fsc` (`module.f.mjs`) | `js/tokenizer` (`module.f.ts`) |
 |---|---|---|
 | continuation | `ToResult = (cp) => [string[], ToResult]` | `ToToken = (cp) => [List<JsToken>, TokenizerState]` |
-| conflict merge | `union` :35 | `union` :330 |
-| `range_map.merge` wrapper | `reduce` :43-51 | `rangeMapMerge` :338-344 |
-| single-range cell | `codePointRange`/`range` :53-55 | `rangeFunc` :346-348 |
-| range-set cell | `rangeSet` :57-66 | `rangeSetFunc` :361-370 |
-| build + dispatch | `create` :68-73 | `create` :372-378 |
+| conflict merge | `union` :33 | `union` :330 |
+| `range_map.merge` wrapper | `reduce` :48-62 | `rangeMapMerge` :338-344 |
+| single-range cell | `codePointRange`/`range` :64-66 | `rangeFunc` :346-348 |
+| range-set cell | `rangeSet` :68-84 | `rangeSetFunc` :361-370 |
+| build + dispatch | `create` :86-97 | `create` :372-378 |
 
 Both `create` functions even end with the byte-identical dispatch shape
 `v => c => x(c)(i)(v)(c)`, where `x = get(def)`.
