+++ fjs/text/todo/190.md
@@ -11,7 +11,7 @@ directly and three of them re-bind the same function under a local name:
 // fjs/media/html/module.f.ts:16
 const { fromCharCode } = String
 
-// fjs/fsc/module.f.ts:19
+// fjs/fsc/module.f.mjs:17
 const fromCharCode = String.fromCharCode
 
 // fjs/js/tokenizer/module.f.ts:73
